Use screen:frame instead of fullFrame
I prefer to have 4 windows in a grid and unfortunately `fullFrame` results in overlapping windows because it doesn't take into account the size of the menu bar at the top of the screen.

Using `frame` fixes this issue because it accounts for the menu bar and dock (which I have hidden, therefore is irrelevant).